CDAP-5461 move parquet plugins back to core-plugins
Undoing temporary changes that were made to get spark-plugins up
and running. Parquet plugins now work when used in the same phase
as a spark plugin.
