+++ core-plugins/docs/SnapshotParquet-batchsink.md
