+++ core-plugins/docs/SnapshotParquet-batchsource.md
