+++ core-plugins/src/main/java/co/cask/hydrator/plugin/batch/source/TimePartitionedFileSetDatasetParquetSource.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.hydrator.plugin.batch.file.source;
+package co.cask.hydrator.plugin.batch.source;
 
 import co.cask.cdap.api.annotation.Description;
 import co.cask.cdap.api.annotation.Name;
@@ -29,7 +29,7 @@ import co.cask.cdap.etl.api.Emitter;
 import co.cask.cdap.etl.api.PipelineConfigurer;
 import co.cask.cdap.etl.api.batch.BatchSource;
 import co.cask.hydrator.common.HiveSchemaConverter;
-import co.cask.hydrator.plugin.batch.file.AvroToStructuredTransformer;
+import co.cask.hydrator.plugin.common.AvroToStructuredTransformer;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.base.Throwables;
