+++ core-plugins/src/test/java/co/cask/hydrator/plugin/batch/ETLStreamConversionTestRun.java
@@ -70,6 +70,11 @@ public class ETLStreamConversionTestRun extends ETLBatchTestBase {
     Schema.Field.of("price", Schema.of(Schema.Type.DOUBLE)));
 
   @Test
+  public void testStreamConversionTPFSParquetSink() throws Exception {
+    testSink("TPFSParquet");
+  }
+
+  @Test
   public void testStreamConversionTPFSAvroSink() throws Exception {
     testSink("TPFSAvro");
   }
