+++ core-plugins/widgets/S3Parquet-batchsink.json
