+++ core-plugins/widgets/SnapshotParquet-batchsink.json
