+++ core-plugins/widgets/SnapshotParquet-batchsource.json
