--- fileset-plugins-parquet/src/test/resources/logback-test.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ Copyright © 2016 Cask Data, Inc.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License"); you may not
-  ~ use this file except in compliance with the License. You may obtain a copy of
-  ~ the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-  ~ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-  ~ License for the specific language governing permissions and limitations under
-  ~ the License.
-  -->
-
-<!--                                                                         -->
-<!-- AppFabric Logging Configuration                                         -->
-<!--                                                                         -->
-<!-- We use the LogBack project for logging in AppFabric. The manual for     -->
-<!-- Logback can be found here: http://logback.qos.ch/manual                 -->
-<!--                                                                         -->
-
-<configuration>
-  <!--Supressing some chatty loggers -->
-  <logger name="org.apache.hadoop" level="INFO"/>
-  <logger name="org.mortbay.log" level="INFO"/>
-
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>%d{ISO8601} - %-5p [%t:%c{1}@%L] - %m%n</pattern>
-    </encoder>
-  </appender>
-
-  <appender name="CLOG" class="co.cask.cdap.common.logging.logback.CAppender">              
-    <layout>                      
-      <!-- Note: we don't insert new line in the pattern as this is done on higher level (in old back-end) -->
-      <pattern>%d{ISO8601} - %-5p [%t:%c{1}@%L] - %m</pattern>       
-    </layout>
-  </appender>
-
-  <logger name="co.cask.hydrator" level="INFO" />
-
-  <root level="WARN">
-    <appender-ref ref="CLOG"/>
-    <appender-ref ref="STDOUT"/>
-  </root>
-
-</configuration>
